Non-zero error exit code, history error output.
Updated the `history` command to emit the error if one is returned to
the callback to `schema.all`.

Called to `process.exit` now include a non-zero exit status when the
program is terminating early due to an exception. This is going to make
it a lot easier to use mygrate with utilities that stop on error, like
`make`.